Token authorization added

+++ handlers/routeHandlers/tokenHandler.js
@@ -7,7 +7,7 @@
 
 //Dependenies
 const libStorage = require("../../lib/data")
-const { hashPassword, parseJson, createRandomString } = require('../../helpers/utilityHelper')
+const { hashPassword, parseJson, createRandomString , verifyToken} = require('../../helpers/utilityHelper')
 const userDefaultDataDirectory = require('../../handlers/routeHandlers/userHandler').defaultDataDirectory
 
 
@@ -28,6 +28,7 @@ handler.tokenHandler = (requestPayload, callback) => {
     return _token[requestPayload.method](requestPayload, callback)
 }
 
+//Login
 _token.post = (requestPayload, callback) => {
     const phone = validatePayload(requestPayload.data.phone, 'string')
     const password = validatePayload(requestPayload.data.password, 'string', 6) 
@@ -53,7 +54,7 @@ _token.post = (requestPayload, callback) => {
         }
 
         const tokenId = createRandomString(20)
-        const expiredAt = Date.now() + 60 * 60 * 1000
+        const expiredAt = Date.now() + 1 * 60 * 1000
         const tokenPayload = {phone, id: tokenId, expiredAt}
         libStorage.create(`${handler.defaultDataDirectory}/token-${tokenId}`, tokenPayload, (err) => {
             if(err) {
@@ -69,27 +70,29 @@ _token.post = (requestPayload, callback) => {
     
     })
 }
-_token.get = (requestPayload, callback) => {
-    const id = validatePayload(requestPayload.url.searchParams.get('id'), 'string')
+// _token.get = (requestPayload, callback) => {
+//     const id = validatePayload(requestPayload.url.searchParams.get('id'), 'string')
 
-    if (!id) {
-        return callback(400, { 
-            message : 'Bad request'
-        })
-    }
+//     if (!id) {
+//         return callback(400, { 
+//             message : 'Bad request'
+//         })
+//     }
 
-    libStorage.get(`${handler.defaultDataDirectory}/token-${id}`, (err, data) => {
-        if(err) {
-            return callback(404, { 
-                message : 'Token not found'
-            })
-        }
-        data = JSON.parse(data)
-        return callback(200, {
-            payload : data
-        })
-    })
-}
+//     libStorage.get(`${handler.defaultDataDirectory}/token-${id}`, (err, data) => {
+//         if(err) {
+//             return callback(404, { 
+//                 message : 'Token not found'
+//             })
+//         }
+//         data = JSON.parse(data)
+//         return callback(200, {
+//             payload : data
+//         })
+//     })
+// }
+
+//Refresh Token
 _token.put = (requestPayload, callback) => {
 
     const id = validatePayload(requestPayload.url.searchParams.get('id'), 'string')
@@ -130,8 +133,10 @@ _token.put = (requestPayload, callback) => {
     })
 
 }
+
+//Logout
 _token.delete = (requestPayload, callback) => {
-    const id = validatePayload(requestPayload.url.searchParams.get('id'), 'string')
+    const id = validatePayload(requestPayload.headers._token, 'string')
 
     if (!id) {
         return callback(400, { 
@@ -142,21 +147,30 @@ _token.delete = (requestPayload, callback) => {
     libStorage.get(`${handler.defaultDataDirectory}/token-${id}`, (err, data) => {
         if(err) {
             return callback(404, { 
-                message : 'Token not found'
+                message : 'Invalid Token'
             })
         }
-
-        libStorage.delete(`${handler.defaultDataDirectory}/token-${id}`, (err) => {
-            if(err) {
-                return callback(500, { 
-                    message : 'Unable to delete'
+        data = JSON.parse(data)
+        verifyToken(id, data.phone, (isTokenVerified) => {
+            if (!isTokenVerified) {
+                return callback(403, { 
+                    message : 'Token expired'
                 })
             }
 
-            return callback(200, {
-                message : 'Token deleted'
+            libStorage.delete(`${handler.defaultDataDirectory}/token-${id}`, (err) => {
+                if(err) {
+                    return callback(500, { 
+                        message : 'Unable to delete'
+                    })
+                }
+    
+                return callback(200, {
+                    message : 'Token deleted'
+                })
             })
         })
+
     })
 }
